server/requirements.txt
microsoft-agents-hosting-aiohttp
cryptography
microsoft-agents-authentication-msal